+++ database/migrations/2026_07_23_135924_remove_self_as_default_type_in_assessment_rater_table.php
@@ -1,0 +1,27 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    public function up(): void
+    {
+        Schema::table('assessment_rater', function (Blueprint $table) {
+            $table->string('type', 10)
+                ->nullable(false)
+                ->change();
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('assessment_rater', function (Blueprint $table) {
+            $table->string('type', 10)
+                ->default('self')
+                ->nullable(false)
+                ->change();
+        });
+    }
+};